% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_onsv.R
\name{tab_onsv}
\alias{tab_onsv}
\title{Add styles to a \code{{gt}} table.}
\usage{
tab_onsv(data)
}
\arguments{
\item{data}{A \code{gt_tbl} object.}
}
\value{
A \code{gt_tbl} customized with \code{tab_style()} and \code{tab_options()}.
}
\description{
\code{tab_onsv()} uses \verb{gt::tab_*()} functions to customize a \code{gt_tbl} object.
}
\examples{
df <- data.frame(
  var1 = c("lorem", "ipsum", "dolor"),
  var2 = c(1, 2, 3)
)

tbl_onsv <-
  df |>
  gt::gt() |>
  tab_onsv()
}
